pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyxctsk"
version = "0.1.0"
description = "Python implementation of XCTrack's task format"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Simon Steiner", email = "simonsteiner@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "geopy>=2.0.0",
    "Pillow>=8.0.0",
    "polyline>=2.0.0",
    "pyproj>=3.0.0",
    "pyzbar>=0.1.9",
    "qrcode[pil]>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "beautifulsoup4>=4.9.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "isort>=5.0.0",
    "mypy>=0.800",
    "pytest-cov>=2.0.0",
    "pytest>=6.0.0",
    "requests>=2.25.0",
]

[project.scripts]
pyxctsk = "pyxctsk.cli:main"

[project.urls]
Homepage = "https://github.com/simonsteiner/pyxctsk"
Repository = "https://github.com/simonsteiner/pyxctsk"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
